migration of arms and API

+++ db/migrate/20190604090134_create_characters.rb
@@ -2,12 +2,12 @@ class CreateCharacters < ActiveRecord::Migration[5.2]
   def change
     create_table :characters do |t|
       t.string :name
-      t.integer :current_life, default: 100
-      t.integer :max_life, default: 100
-      t.integer :power, default: 2
+      t.float :current_life, default: 100.0
+      t.float :max_life, default: 100.0
+      t.float :power, default: 2.0
       t.integer :level, default: 1
-      t.integer :current_experience, default: 0
-      t.integer :max_experience, default: 100
+      t.float :current_experience, default: 0.0
+      t.float :max_experience, default: 100.0
 
       t.timestamps null: false
     end
