+++ db/migrate/20190605090131_create_fightings.rb
@@ -4,7 +4,7 @@ class CreateFightings < ActiveRecord::Migration[5.2]
       t.references :fighter_one, index: true, foreign_key: { to_table: :characters }
       t.references :fighter_two, index: true, foreign_key: { to_table: :characters }
       t.string :winner
-      t.integer :winner_gained_experience, default: 0
+      t.float :winner_gained_experience, default: 0.0
       t.boolean :winner_leveled_up, default: false
 
       t.timestamps
